% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddNorthArrow.R
\name{AddNorthArrow}
\alias{AddNorthArrow}
\title{Plot north arrow}
\source{
<https://CRAN.R-project.org/package=prettymapr>
}
\usage{
AddNorthArrow(
  pos = "topright",
  padin = c(0.15, 0.15),
  scale = 1,
  lwd = 1,
  border = "black",
  cols = c("white", "black"),
  text.col = "black"
)
}
\arguments{
\item{pos}{where to align the north arrow. One of "bottomleft",
"bottomright", "topleft", or "topright".}

\item{padin}{a vector of length 2 determining the distance in inches between
the scalebar and the edge of the plottable area.}

\item{scale}{scale the default north arrow to make it bigger or smaller.}

\item{lwd}{the line width outlining the north arrow.}

\item{border}{the line color outlining the north arrow.}

\item{cols}{a vector of length 2 determining the two colors to be drawn for
the north arrow.}

\item{text.col}{color of the "N".}
}
\description{
Plot a north arrow (pointing directly "up") positioned based on current plot
extents.
}
\examples{

plot(1 : 5, 1 : 5, asp = 1)
AddNorthArrow()

}
\author{
Dewey Dunnington
}
