% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabelAxis.R
\name{bquote.font}
\alias{bquote.font}
\title{Font selection for bquote}
\usage{
bquote.font(obj, font = 1)
}
\arguments{
\item{obj}{a language object.}

\item{font}{an integer which specifies which font to use for text; see
\code{?par} for the available options.}
}
\value{
a language object.
}
\description{
A wrapper function for \code{\link[base]{bquote}} to select a text font.
}
\author{
Thomas Münch
}
