% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColorPal.R
\name{ColorPal}
\alias{ColorPal}
\title{ColorBrewer colour palettes}
\source{
The ColorBrewer 2.0 collection can be viewed interactively at
  https://colorbrewer2.org.
}
\usage{
ColorPal(name = "RdYlBu", n = 10, rev = FALSE, fun = FALSE)
}
\arguments{
\item{name}{a palette name from the ColorBrewer 2.0 collection; defaults to
the diverging \code{RdYlBu} palette.}

\item{n}{integer number of different colours in the palette, minimum 3,
maximum depending on the chosen palette. Default is \code{10} for the
\code{RdYlBu} palette.}

\item{rev}{logical; set to \code{TRUE} to reverse the order of colours in the
palette.}

\item{fun}{logical to control the return type; if set to \code{FALSE} (the
default) returned is simply a character vector of length \code{n} with the
hexadecimal colour codes from the requested palette. Set to \code{TRUE} to
instead obtain a function that takes an integer argument (the required
number of colours) and returns a character vector of hexadecimal colour
codes, which is needed, e.g., for a filled contour plot
(\code{\link[graphics]{filled.contour}}).}
}
\value{
either a character vector of hexadecimal colour codes, or a function
  to return a colour code vector.
}
\description{
Obtain a colour palette from the ColorBrewer 2.0 collection for use in
standard plots, image plots, or filled contour plots.
}
\examples{

# default palette
ColorPal()

# reverse the order of colours
ColorPal(rev = TRUE)

# return a colour palette function instead for use in filled contour plots
fun <- ColorPal(fun = TRUE)
fun(5)

}
\seealso{
\code{\link[grDevices]{colorRampPalette}};
  \code{\link[RColorBrewer]{brewer.pal}}
}
\author{
Thomas Münch
}
