% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPairwiseCorrelations.R
\name{PlotPairwiseCorrelations}
\alias{PlotPairwiseCorrelations}
\title{Calculate and plot a pairwise correlations}
\usage{
PlotPairwiseCorrelations(
  M,
  axis.label = NULL,
  return.corr.tibble = FALSE,
  plotit = TRUE
)
}
\arguments{
\item{M}{A matrix or dataframe for which the pairwise corrlelations between
the columns are desired.}

\item{axis.label}{Optional label for x,y figure axes, Default: NULL}

\item{return.corr.tibble}{Return the correlations?, Default: FALSE}

\item{plotit}{Plot the correlations, Default: TRUE}
}
\value{
Default: a ggplot2 object, optionally the correlations as a tibble.
}
\description{
Takes a matrix or dataframe and calculates the pairwise
  correlations between all columns. Plots these as a ggplot tilemap.
  Optionally returns the pairwise correlations as a tidy tibble.
}
\details{
Ideas for code taken from here:
  http://www.sthda.com/english/wiki/ggplot2-quick-correlation-matrix-heatmap-r-software-and-data-visualization
}
\examples{
\dontrun{
if(interactive()) {

  M <- matrix(rnorm(12 * 100), ncol = 12)
  PlotPairwiseCorrelations(M, axis.label = "Month",
                           return.corr.tibble = TRUE)

  # Add names to columns
  colnames(M) <- letters[1:12]
  PlotPairwiseCorrelations(M, axis.label = "Cat",
                           return.corr.tibble = TRUE)

}
}
}
\author{
Andrew Dolman
}
