% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErrorBars.R
\name{ErrorBars}
\alias{ErrorBars}
\title{Add error bars}
\usage{
ErrorBars(
  x,
  y,
  upper,
  lower = upper,
  width = 0.05,
  col = "black",
  lwd = 1,
  lty = 1
)
}
\arguments{
\item{x}{x coordinates of the points for which error bars are to be added.}

\item{y}{y coordinates of the points for which error bars are to be added.}

\item{upper}{upper value of the error interval relative to \code{y}; i.e.,
the upper end of the error interval extends to \code{y + upper}.}

\item{lower}{same as \code{upper} for the lower value of the error interval;
default is to assume symmetric error intervals around \code{y}.}

\item{width}{the width of the error bar drawing as a fraction of the x user
coordinate range (\code{diff(range(par("usr")[1 : 2]))}); defaults to five
percent.}

\item{col}{the colour of the error bars; default "black".}

\item{lwd}{the line width of the error bars; default \code{1}.}

\item{lty}{the line type of the error bars; default are solid lines
(\code{lty = 1}).}
}
\description{
Add error bars to an existing point plot.
}
\examples{
x <- 1 : 10
plot(x, ylim = c(0, 12))
ErrorBars(x, y = x, upper = 2, lower = 1)
ErrorBars(x, y = x, upper = 0.5, col = "red", width = 0.025)
}
\author{
Thomas Münch
}
