% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapPlotly.R
\name{MapPlotly}
\alias{MapPlotly}
\title{MapPlotly}
\usage{
MapPlotly(
  proj = c("equirectangular", "mercator", "orthographic", "natural earth", "kavrayskiy7",
    "miller", "robinson", "eckert4", "azimuthal equal area", "azimuthal equidistant",
    "conic equal area", "conic conformal", "conic equidistant", "gnomonic",
    "stereographic", "mollweide", "hammer", "transverse mercator", "winkel tripel",
    "aitoff", "sinusoidal"),
  resolution = 110,
  focal.lat = 0,
  focal.lon = 0,
  lon.for.latlabels = -30,
  lat.for.lonlabels = 15,
  lon.lines = seq(-180, 180, 45),
  lat.lines = seq(-90, 90, 30),
  lat.range = c(-90, 90),
  lon.range = c(0, 360),
  show.longrid = TRUE,
  show.latgrid = TRUE,
  land.color = "#e5ecf6",
  ocean.color = "white",
  colors = "Set2"
)
}
\arguments{
\item{proj}{name of the projection to use}

\item{resolution}{resolution of the base map, options are 110 or 50 km per mm. Defaults to 110 km/mm}

\item{focal.lat}{centre on this latitude}

\item{focal.lon}{centre on this longitude}

\item{lon.for.latlabels}{put a line of latitude labels along this longitude (can use multiple)}

\item{lat.for.lonlabels}{put a line of longitude labels along this latitude (can use multiple)}

\item{lon.lines}{plot these longitude lines as a graticule}

\item{lat.lines}{plot these latitude lines as a graticule}

\item{lat.range}{restrict the plot to this latitude range e.g. c(-30, 30)}

\item{lon.range}{restrict the plot to this longitude range e.g. c(-30, 90)}

\item{show.longrid, show.latgrid}{show fine gridlines}

\item{land.color}{fill colour for land}

\item{ocean.color}{fill colour of ocean}

\item{colors}{colour palette passed to \code{plotly::plot_geo}}
}
\value{
a plotly widget
}
\description{
Create interactive maps (plotly htmlwidgets) with many different projections.
}
\examples{
\dontrun{
# Mollenweide
p <- MapPlotly(proj = "moll")
p

# Mollenweide with North Pole focus
p <- MapPlotly(proj = "moll", focal.lat = 90)
p

p <- MapPlotly(proj = "azimuthal equal area", focal.lat = 90)
p

# Not so great for whole globe, but limit lat

p <- MapPlotly(proj = "azimuthal equal area", focal.lat = 90, lat.range  = c(0, 90))
p

p <- MapPlotly(proj = "azimuthal equal area", focal.lat = 0, lon.range = c(-60, 60))
p 

# basemap a little low res for small areas
p <- MapPlotly(
 proj = "azimuthal equal area", focal.lat = 53, lon.range = c(-35, 35), lat.range = c(30, 60) )
p

# switch off labels for some things
p <- MapPlotly(proj = "moll", focal.lat = 0, focal.lon = 0, 
          lon.for.latlabels = NA, lat.for.lonlabels = NA,
          lon.lines = NA, lat.lines = NA,
          show.latgrid = FALSE, show.longrid = FALSE)
p


# Add datapoints (see plotly for more things like this)

dat <- data.frame(lat = c(15, 45), lon = c(22, -90),
              type = c("Pollen", "Tree ring"))

plotly::add_markers(p, data = dat, showlegend = TRUE,
              color = I("blue"))



p <- MapPlotly(proj = "ortho", focal.lat = 90, focal.lon = 135,
          # lon.lines = seq(-80, -10, 10),
          # lon.for.latlabels = -55, lat.for.lonlabels = 65,
          lat.range = c(0, 90))
p <- plotly::add_markers(p, data = dat, showlegend = TRUE,
              color = I("blue"))
p

p <- MapPlotly(proj = "ortho", focal.lat = 0, focal.lon = 0,
               lon.range = c(0, 360),
               lat.range = c(-90, 90))  

p

p <- MapPlotly(proj = "eckert4",
               focal.lat = 0, focal.lon = -90          )
p

p <- MapPlotly(proj = "eckert4",
               focal.lat = 0, focal.lon = -90, lat.range = c(-40, 40), lon.range = c(45, 170))
p


p <- MapPlotly(proj = "eckert4",
               focal.lat = 90, focal.lon = 135)
p
}
}
