% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quartz.R
\name{Par}
\alias{Par}
\title{Set new graphical parameters}
\usage{
Par(
  mar = c(5, 5, 0.5, 0.5),
  las = 1,
  cex.axis = 1.25,
  cex.lab = 1.5,
  cex.main = 1.5,
  ...
)
}
\arguments{
\item{mar}{a numerical vector of the form ‘c(bottom, left, top, right)’ which
gives the number of lines of margin to be specified on the four sides of
the plot. The default is \code{c(5, 5, 0.5, 0.5)}.}

\item{las}{the style of axis labels; default is to always use horizontal axis
labels (note that you then need to manually set \code{las = 0} for the y
axis titles when using \code{mtext}).}

\item{cex.axis}{the magnification to be used for axis annotation relative to
the current setting of \code{cex}; defaults to \code{1.25}.}

\item{cex.lab}{the magnification to be used for x and y labels relative to the
current setting of \code{cex}; defaults to \code{1.5}.}

\item{cex.main}{the magnification to be used for main titles relative to the
current setting of \code{cex}; defaults to \code{1.5}.}

\item{...}{further graphical parameter settings passed on to
\code{\link[graphics]{par}}.}
}
\value{
invisibly, the original graphical parameter settings prior to the
  changes from a call to \code{Par} to enable restoring the default
  settings; see the example.
}
\description{
This function sets the graphical parameters which are specified as its
function arguments in a call to \code{\link[graphics]{par}}. The function's
default settings are set to produce visually more appealing plots compared to
base R's default settings. You can change the function's defaults to your
needs and/or set additional graphical parameters by passing them in \code{tag
= value} form.
}
\examples{
op <- Par()
plot(1 : 10, type = "l", main = "", xlab = "X title", ylab = "Y title")
par(op)
}
\seealso{
\code{\link[graphics]{par}}
}
\author{
Thomas Münch
}
