% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotStamp.R
\name{PlotStamp}
\alias{PlotStamp}
\title{Add session information to plot}
\usage{
PlotStamp(script, add.info = NULL, incr = 0.01, cex = 0.75)
}
\arguments{
\item{script}{character string with the name of the script used for creating
the plot.}

\item{add.info}{character string with additional information to be added
(optional).}

\item{incr}{numeric; fraction of the user coordinate range to use as
incremental distance from the default position of the information text in
the upper right corner of the plot region.}

\item{cex}{the magnification to be used for the information text relative to
the current setting of \code{cex}; defaults to \code{0.75}.}
}
\description{
Add session information to an existing base plot: script name, date of plot
creation, etc., placed in the upper right corner of the plot.
}
\examples{
plot(1 : 10, type = "l")
PlotStamp(script = "interactive",
          add.info = "Illustrated usage of PlotStamp function")
}
\author{
Thomas Münch
}
