% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Polyplot.R
\name{Polyplot}
\alias{Polyplot}
\title{Draw error shading}
\usage{
Polyplot(x, y1, y2, col = "black", alpha = 0.2, ...)
}
\arguments{
\item{x}{numeric vector of x values of the error band.}

\item{y1}{numeric vector for the upper bound of the error band; must be of
the same length as \code{x}.}

\item{y2}{numeric vector for the lower bound of the error band; must be of
the same length as \code{x}.}

\item{col}{colour of the error band.}

\item{alpha}{opacity factor for \code{col} within [0,1].}

\item{...}{additional parameters which are passed to \code{polygon}.}
}
\description{
A wrapper function for the \code{polygon} function to draw error shadings
(or confidence intervals) on a line plot.
}
\examples{
x <- 1 : 10
plot(x, type = "n", xlab = "x", ylab = "y")
Polyplot(x, y1 = x + 2, y2 = x - 2)
lines(x, lwd = 2)
}
\seealso{
\code{\link{polygon}}
}
\author{
Thomas Münch
}
