% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapPlotly.R
\name{SetupPlotlySave}
\alias{SetupPlotlySave}
\title{SetupPlotlySave}
\usage{
SetupPlotlySave()
}
\description{
Install the Python kaleido package needed to save plotly figures
  as static images via \code{SaveMapPlotly}. Only needs to be run once per
  machine. Requires the \pkg{reticulate} package.
}
\examples{
\dontrun{
SetupPlotlySave()
}
}
