% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddNorthArrow.R
\name{AddNorthArrow}
\alias{AddNorthArrow}
\title{Plot north arrow}
\source{
Adapted from
  <https://CRAN.R-project.org/package=prettymapr>
}
\usage{
AddNorthArrow(
  pos = "topright",
  padin = c(0.15, 0.15),
  scale = 1,
  lwd = 1,
  border = "black",
  cols = c("white", "black"),
  col.text = "black",
  cex.text = 1,
  vadj.text = 0.5
)
}
\arguments{
\item{pos}{where to align the north arrow. One of "bottomleft",
"bottomright", "topleft", or "topright".}

\item{padin}{a vector of length 2 determining the distance in inches between
the scalebar and the edge of the plottable area.}

\item{scale}{scale the north arrow to make it bigger or smaller.}

\item{lwd}{the line width outlining the north arrow.}

\item{border}{the line color outlining the north arrow.}

\item{cols}{a vector of length 2 determining the two colors to be drawn for
the north arrow.}

\item{col.text}{color of the text label.}

\item{cex.text}{character expansion factor for the text label. This is an
absolute measure, not scaled by \code{par("cex")} or by setting
\code{par("mfrow")} or \code{par("mfcol")}.}

\item{vadj.text}{value in [0, 1] to specify the vertical adjustment of the
text label relative to the north arrow, with 0 for top, 1 for bottom, and
0.5 for centered position. On most devices values outside [0, 1] will also
work.}
}
\description{
Plot a north arrow (pointing directly "up" and with label "N") positioned
based on current plot extents.
}
\examples{

plot(1 : 5, 1 : 5, asp = 1)
AddNorthArrow()

}
\author{
Dewey Dunnington, Thomas Münch
}
