% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-helpers.R
\name{facet_wrap_paginate_auto}
\alias{facet_wrap_paginate_auto}
\title{Plot all ggforce::facet_wrap_paginate pages}
\usage{
facet_wrap_paginate_auto(ggplot.obj, facets, nrow, ncol, scales = "free")
}
\arguments{
\item{ggplot.obj}{a ggplot2 object.}

\item{facets}{faceting formula; see \code{\link[ggplot2]{facet_wrap}}.}

\item{nrow}{number of columns per page.}

\item{ncol}{number of rows per page.}
}
\value{
A list of ggplot2 objects.
}
\description{
A helper function to return all pages from ggforce::facet_wrap_paginate. It
automatically calculates the required number of pages.
}
\examples{
library(ggplot2)
dat <- expand.grid(x = 1:100, a = letters[1:11])
dat$y <- dat$x + rnorm(nrow(dat), 0, 10)
p <- ggplot(dat, aes(x = x, y = y)) +
  geom_point() +
  facet_wrap(~a)
p

gg <- facet_wrap_paginate_auto(p,
                               facets = formula(~a),
                               nrow = 2, ncol = 2)
gg
}
