% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quartz.R
\name{Quartz}
\alias{Quartz}
\title{Open quartz device}
\usage{
Quartz(
  file = NULL,
  type = "native",
  height = 6,
  width = 8,
  mar = c(5, 5, 0.5, 0.5),
  las = 1,
  cex.axis = 1.25,
  cex.lab = 1.5,
  cex.main = 1.5,
  ...,
  pointsize = 12,
  family = "Arial",
  antialias = TRUE,
  bg = "transparent",
  canvas = "white",
  dpi = NA_real_
)
}
\arguments{
\item{file}{path to a file for storing a hardcopy of the plot including a
supported file extension to set the \code{type} of output (e.g. ".pdf" or
".png"); i.e. the function extracts the extension from \code{file} and
uses it as the \code{type} argument for the call to
\code{\link[grDevices]{quartz}}. Defaults to \code{NULL} for on-screen
plotting.}

\item{type}{the type of output to use. Defaults to \code{"native"} for
on-screen plotting. If \code{file} is not \code{NULL}, \code{type} is
determined from the file name extension.}

\item{height}{the height of the plotting area in inches.  Default ‘6’.}

\item{width}{the width of the plotting area in inches.  Default ‘8’.}

\item{mar}{a numerical vector of the form ‘c(bottom, left, top, right)’ which
gives the number of lines of margin to be specified on the four sides of
the plot. The default is \code{c(5, 5, 0.5, 0.5)}.}

\item{las}{the style of axis labels; default is to always use horizontal axis
labels (note that you then need to manually set \code{las = 0} for the y
axis titles when using \code{mtext}).}

\item{cex.axis}{the magnification to be used for axis annotation relative to
the current setting of \code{cex}; defaults to \code{1.25}.}

\item{cex.lab}{the magnification to be used for x and y labels relative to the
current setting of \code{cex}; defaults to \code{1.5}.}

\item{cex.main}{the magnification to be used for main titles relative to the
current setting of \code{cex}; defaults to \code{1.5}.}

\item{...}{further graphical parameter settings passed on to
\code{\link{Par}}.}

\item{pointsize}{the default pointsize to be used.  Default \code{12}.}

\item{family}{this is the family name of the font
    that will be used by the device.  Default \code{"Arial"}.  This will
    be the base name of a font as  shown in Font Book.}

\item{antialias}{whether to use antialiasing.  Default \code{TRUE}.}

\item{bg}{the initial background colour to use for the device.  Default
    \code{"transparent"}.  An opaque colour such as \code{"white"} will
    normally be required on off-screen types that support transparency
    such as \code{"png"} and \code{"tiff"}.}

\item{canvas}{canvas colour to use for an on-screen device.  Default
    \code{"white"}, and will be forced to be an opaque colour.}

\item{dpi}{resolution of the output. The default (‘NA_real_’) for an
on-screen display defaults to the resolution of the main screen, and to 72
dpi otherwise.}
}
\value{
invisibly, the original graphical parameter settings prior to the
  changes from a call to \code{Quartz} to enable restoring the default
  settings.
}
\description{
Wrapper to open a quartz device with default 4:3 dimensions for on-screen
plotting or saving to a file, including the setting of default graphical
parameters which are deliberately different from base R's setting to produce
visually more appealing plots.
}
\examples{

# Create an empty on-screen quartz device
Quartz()

# Store empty plot in pdf format in local directory
\dontrun{
Quartz(file = file.path(getwd(), "test-quartz.pdf"))
dev.off()
}
}
\seealso{
\code{\link[grDevices]{quartz}}; \code{\link{Par}}
}
\author{
Thomas Münch
}
